multi_split_prec: fixed bug for coarse M != 0
The origin of the problem was the difference in results between pure
diagonal and degree cases when we had M+Md=N on the coarsest level. The
problem was that we set the number of iterations on the coarsest level
to be one, but that only works when we solve the coarse system directly,
i.e. when M = 0 and Md = N. When we started using other elimination
techniques, we got this issue.