Fixed default reading of the SPE mesh
The SPE tensor data downloaded from the website is actually a text file,
not a binary one. Thus, I needede to modify the reading routine.

In the long term (and also in the past), the file should be converted to
a binary format to speed up its reading, and the routine should read it
in binarily.